+++ angular-client/src/pages/lap-timer-page/laps-table/laps-table.component.css
@@ -1,0 +1,114 @@
+.monospace {
+  font-family: var(--font-family);
+  font-variant-numeric: tabular-nums;
+}
+
+.lap-table-wrap {
+  flex: 1 1 0;
+  min-height: 0;
+  width: 100%;
+  overflow-y: auto;
+  overflow-x: hidden;
+  scrollbar-width: thin;
+  scrollbar-color: var(--color-divider) transparent;
+}
+.lap-table-wrap::-webkit-scrollbar {
+  display: block;
+  width: 6px;
+}
+.lap-table-wrap::-webkit-scrollbar-thumb {
+  background: var(--color-divider);
+  border-radius: 3px;
+}
+.lap-table-wrap::-webkit-scrollbar-track {
+  background: transparent;
+}
+
+.empty-laps {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 32px;
+}
+
+.lap-table-wrap :where(.p-datatable) {
+  font-family: var(--font-family);
+  font-size: var(--font-size-sm);
+  color: var(--color-text-primary);
+}
+.lap-table-wrap :where(th) {
+  font-size: var(--font-size-xs);
+  font-weight: 600;
+  letter-spacing: 0.5px;
+  color: var(--color-text-secondary);
+  text-transform: uppercase;
+  background: transparent;
+  border-bottom: 1px solid var(--color-divider);
+  padding: 8px 12px;
+  text-align: left;
+}
+.lap-table-wrap :where(td) {
+  border-bottom: 1px solid var(--color-divider);
+  padding: 8px 12px;
+  vertical-align: middle;
+}
+.lap-table-wrap :where(tr):hover td {
+  background: rgba(255, 255, 255, 0.03);
+}
+
+.lap-row-best > td:first-child {
+  border-left: 3px solid var(--color-battery-high);
+}
+.lap-row-worst > td:first-child {
+  border-left: 3px solid var(--color-battery-low);
+}
+
+/* !important: overrides PrimeNG's higher-specificity th/td defaults. */
+.lap-table-wrap th.col-num,
+.lap-table-wrap td.col-num {
+  width: 56px;
+  text-align: right !important;
+}
+.lap-table-wrap th.col-dur,
+.lap-table-wrap td.col-dur {
+  width: 100px;
+  text-align: right !important;
+}
+.lap-table-wrap th.col-delta,
+.lap-table-wrap td.col-delta {
+  width: 110px;
+  text-align: right !important;
+}
+.lap-table-wrap th.col-time,
+.lap-table-wrap td.col-time {
+  width: 130px;
+  text-align: right !important;
+}
+.lap-table-wrap th.col-run,
+.lap-table-wrap td.col-run {
+  width: 64px;
+  text-align: right !important;
+}
+.lap-table-wrap th.col-stat,
+.lap-table-wrap td.col-stat {
+  width: 100px;
+  text-align: right !important;
+}
+
+.delta-positive {
+  color: var(--color-battery-low);
+}
+
+.best-flag {
+  display: inline-flex;
+  align-items: center;
+  gap: 4px;
+  color: var(--color-battery-high);
+  font-weight: 600;
+}
+.best-flag mat-icon {
+  font-size: 16px;
+  width: 16px;
+  height: 16px;
+}
