+++ angular-client/src/pages/lap-timer-page/timer-hero/timer-hero.component.html
@@ -1,0 +1,74 @@
+<info-background
+  svgIcon="timer"
+  title="Lap Timer"
+  [topRightInfo]="timer.activeSession()?.name"
+  style="width: 100%; height: 100%"
+>
+  <div class="timer-hero">
+    <div class="hero-label">TOTAL</div>
+    <div class="hero-total">{{ timer.formattedTotal() }}</div>
+    <div class="hero-divider"></div>
+    <div class="hero-current-row">
+      <div class="hero-label">LAP {{ timer.lapCount() + 1 }}</div>
+      <div class="hero-current-time">{{ timer.formattedCurrentLap() }}</div>
+      <div class="hero-current-delta" [class]="currentLapDeltaClass()">{{ currentLapDeltaText() }}</div>
+    </div>
+    <div class="hero-actions">
+      @if (!timer.activeSession()) {
+        <argos-button
+          label="Start"
+          [onClick]="onStart"
+          additionalStyles="background-color: var(--color-battery-high); color: #000; font-weight: 700;"
+        />
+      } @else if (timer.isIdle()) {
+        <argos-button
+          label="Start"
+          [onClick]="onStart"
+          additionalStyles="background-color: var(--color-battery-high); color: #000; font-weight: 700;"
+        />
+        <argos-button
+          label="Delete"
+          [onClick]="onDeleteActiveButton"
+          additionalStyles="background-color: var(--color-battery-low); color: #fff; font-weight: 600;"
+        />
+      } @else if (timer.isRunning()) {
+        <argos-button
+          label="Pause"
+          [onClick]="onPause"
+          additionalStyles="background-color: var(--color-battery-med); color: #000; font-weight: 700;"
+        />
+        <argos-button
+          label="Lap"
+          [onClick]="onLap"
+          additionalStyles="background-color: var(--color-battery-high); color: #000; font-weight: 700;"
+        />
+        <argos-button
+          label="Stop"
+          [onClick]="onStop"
+          additionalStyles="background-color: var(--color-battery-low); color: #fff; font-weight: 600;"
+        />
+      } @else if (timer.isPaused()) {
+        <argos-button
+          label="Resume"
+          [onClick]="onResume"
+          additionalStyles="background-color: var(--color-battery-high); color: #000; font-weight: 700;"
+        />
+        <argos-button
+          label="Reset"
+          [onClick]="onReset"
+          additionalStyles="background-color: var(--color-battery-low); color: #fff; font-weight: 600;"
+        />
+        <argos-button
+          label="Download"
+          [onClick]="onDownloadActiveButton"
+          additionalStyles="background-color: var(--color-background-info); border: 1px solid var(--color-divider); color: var(--color-text-primary);"
+        />
+        <argos-button
+          label="End Session"
+          [onClick]="onEndActiveSession"
+          additionalStyles="background-color: var(--color-background-info); border: 1px solid var(--color-divider); color: var(--color-text-primary);"
+        />
+      }
+    </div>
+  </div>
+</info-background>
